+++ src/components/groupDetail/GroupOwners.vue
@@ -5,7 +5,7 @@ import { ref } from 'vue'
 import { useGroupDetailStore } from '/@/stores/groupDetail'
 import { useUserStore } from '/@/stores/user'
 
-import FormSelect from '/@/components/shared/FormSelect.vue'
+import InputSelect from '/@/components/shared/InputSelect.vue'
 import UserIcon from '/@/components/shared/UserIcon.vue'
 
 import { useGroupOwner } from './composables/useGroupOwner'
@@ -44,7 +44,7 @@ const { absentOwnerOptions, isSending, addOwners, removeOwner } =
     </ul>
     <div v-if="hasAuthority" class="flex p-2">
       <!--todo:https://vue-select.org/guide/positioning.html#default-->
-      <FormSelect
+      <InputSelect
         v-model="OwnersToBeAdded"
         class="mr-2 flex-grow"
         is-multiple
