+++ src/components/requestDetail/RequestTags.vue
@@ -5,7 +5,7 @@ import { useRequestDetailStore } from '/@/stores/requestDetail'
 import { useUserStore } from '/@/stores/user'
 
 import EditButton from '/@/components/shared/EditButton.vue'
-import FormTagSelectWithCreation from '/@/components/shared/FormTagSelectWithCreation.vue'
+import InputSelectTagWithCreation from '/@/components/shared/InputSelectTagWithCreation.vue'
 import SimpleButton from '/@/components/shared/SimpleButton.vue'
 import TagsGroup from '/@/components/shared/TagsGroup.vue'
 import type { EditMode } from '/@/pages/composables/requestDetail/useRequestDetail'
@@ -44,7 +44,7 @@ const handleComplete = () => {
     </div>
     <div v-else class="flex items-center">
       タグ：
-      <FormTagSelectWithCreation v-model="editedValue.tags" />
+      <InputSelectTagWithCreation v-model="editedValue.tags" />
       <SimpleButton
         class="ml-2"
         font-size="sm"
