+++ src/components/requestDetail/RequestTitle.vue
@@ -4,7 +4,7 @@ import { storeToRefs } from 'pinia'
 import { useRequestDetailStore } from '/@/stores/requestDetail'
 import { useUserStore } from '/@/stores/user'
 
-import FormInput from '/@/components/shared/FormInput.vue'
+import InputText from '/@/components/shared/InputText.vue'
 import SimpleButton from '/@/components/shared/SimpleButton.vue'
 import type { EditMode } from '/@/pages/composables/requestDetail/useRequestDetail'
 
@@ -40,7 +40,7 @@ const hasAuthority = requestDetailStore.isRequestCreater(userStore.me)
     </SimpleButton>
   </div>
   <div v-else class="flex flex-grow">
-    <FormInput
+    <InputText
       v-model="editedValue.title"
       class="flex-grow"
       placeholder="タイトル"
